refactor(node-fs): separate files
