+++ packages/node-fs/src/logger.ts
@@ -1,0 +1,5 @@
+import {definePackage} from '@alwatr/logger';
+
+import type {} from '@alwatr/nano-build';
+
+export const logger = definePackage('@alwatr/node-fs', __package_version__);
